+++ source/ftrack_connect/ui/widget/line_edit.py
@@ -1,10 +1,10 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014 ftrack
 
-from PySide import QtGui, QtCore
+from QtExt import QtWidgets, QtCore, QtGui
 
 
-class LineEditIconButton(QtGui.QToolButton):
+class LineEditIconButton(QtWidgets.QToolButton):
     '''Icon button for use in a :py:class:`LineEdit` widget.'''
 
     iconSize = 16
@@ -18,7 +18,7 @@ class LineEditIconButton(QtGui.QToolButton):
 
     def paintEvent(self, event):
         '''Handle paint *event*.'''
-        painter = QtGui.QPainter(self)
+        painter = QtWidgets.QPainter(self)
 
         # Note: isDown should ideally use the 'active' state, but in most styles
         # this has no proper feedback.
@@ -42,7 +42,7 @@ class LineEditIconButton(QtGui.QToolButton):
         painter.drawPixmap(iconRegion, iconPixmap)
 
 
-class LineEdit(QtGui.QLineEdit):
+class LineEdit(QtWidgets.QLineEdit):
     '''Line edit that supports embedded actions.'''
 
     def __init__(self, *args, **kw):